Fix context range precision gap
